+++ app/src/main/AndroidManifest.xml
@@ -10,21 +10,25 @@
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
 
+
         <activity
                 android:name=".MainActivity"
                 android:label="@string/app_name"
                 android:theme="@style/AppTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
+
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-
+        <activity android:name=".MemoListActivity"/>
         <activity android:name=".CreateActivity"/>
         <activity android:name=".ReadActivity"/>
+
         <meta-data
                 android:name="com.google.android.actions"
                 android:resource="@xml/actions"/>
+
 
     </application>
