+++ app/src/main/java/com/example/s_kim/memoproject/ChatInfo.kt
@@ -1,0 +1,9 @@
+package com.example.s_kim.memoproject
+
+
+class ChatInfo(
+    var listNumber: Int,
+    var listTitle: String,
+    var chatMessage: String
+)
+
